+++ lib/tmdb_dart.dart
@@ -1,20 +1,22 @@
-export 'src/settings/configuration.dart';
+export 'src/settings/append-settings.dart';
 export 'src/settings/asset-resolver.dart';
-export 'src/settings/movie-search-settings.dart';
-export 'src/settings/movie-append-settings.dart';
+export 'src/settings/configuration.dart';
 export 'src/settings/quality-settings.dart';
+export 'src/settings/search-settings.dart';
 export 'src/models/paged-result.dart';
 export 'src/models/movie.dart';
 export 'src/models/genre.dart';
 export 'src/models/country.dart';
 export 'src/models/company.dart';
 export 'src/models/collection.dart';
+export 'src/models/content-base.dart';
 export 'src/models/image-info.dart';
 export 'src/models/image-collection.dart';
-export 'src/models/movie-base.dart';
+export 'src/models/tv-show.dart';
 
 export 'src/models/alternative-title.dart';
 export 'src/models/cast.dart';
+export 'src/models/creator.dart';
 export 'src/models/crew.dart';
 export 'src/models/credits.dart';
 export 'src/models/date.dart';